requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
replicate==0.22.0
pillow==10.2.0
python-dotenv==1.0.0
aiofiles==23.2.1
